make some refactoring
