+++ app/controllers/auth.controller.js
@@ -1,4 +1,4 @@
-import { Auth, File } from './../methods/index';
+import { Auth } from './../methods/index';
 
 export default class AuthContoller {
 
@@ -25,41 +25,7 @@ export default class AuthContoller {
     } else {
       responce = 'Invalid Token';
     }
-
     ctx.body = responce;
-    await next();
-  }
-
-  static async testprivate(ctx, next) {
-    ctx.body = 'testprivate';
-    await next();
-  }
-
-  static async testpublic(ctx, next) {
-    ctx.body = 'testpublic';
-    await next();
-  }
-
-  static async fileUpload(ctx, next) {
-    await File.upload(ctx);
-    await next();
-  }
-
-  static async fileDelete(ctx, next) {
-    const { request: { body: { path } } } = ctx;
-    await File.delete(path, true);
-    await next();
-  }
-
-  static async filePath(ctx, next) {
-    const { request: { body: { path } } } = ctx;
-    ctx.body = await File.path(path);
-    await next();
-  }
-
-  static async fileExist(ctx, next) {
-    const { request: { body: { path } } } = ctx;
-    ctx.body = await File.exist(path);
     await next();
   }
